requirements.txt
# Core Flask
Flask==3.0.3
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
click==8.1.7

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
Flask-Login==0.6.3
PyJWT==2.8.0
email-validator==2.1.0.post1

# File Processing
Pillow==11.0.0
pdf2image==1.17.0
pytesseract==0.3.10
python-magic==0.4.27
python-dotenv==1.0.1

# Forms and Templates
Flask-WTF==1.2.1
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.1
WTForms==3.1.2

# Email
Flask-Mail==0.9.1

# Utilities
python-slugify==8.0.1
requests==2.31.0
python-dateutil==2.9.0.post0
six==1.16.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
greenlet==3.0.3

# Development
pytest==8.0.0
pytest-cov==4.1.0
coverage==7.3.2
